3.2-crud/stocks_products/requirements.txt
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
psycopg2-binary==2.9.10